edit profile option, search + category filter option, added deep validation (only user that created their post can edit/delete it), css changes

+++ app/Http/Controllers/PostController.php
@@ -5,22 +5,20 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Models\Post;
 use App\Models\Category;
-use App\Models\ViewedPost;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Support\Facades\Auth;
 
 
 class PostController extends Controller
 {
-
     public function index(Request $request)
     {
         // Retrieve all posts for admin users, including invisible ones
         if (Auth::check() && Auth::user()->role === 'admin') {
-            $posts = Post::get();
+            $query = Post::query();
         } else {
             // For non-admin users, retrieve only visible posts
-            $posts = Post::where('is_visible', true)->get();
+            $query = Post::where('is_visible', true);
         }
 
         // Retrieve all categories
@@ -29,25 +27,66 @@ class PostController extends Controller
         // Retrieve the category ID from the request
         $categoryId = $request->input('category');
 
-        // Initialize the query based on user role
-        $query = Auth::check() && Auth::user()->role === 'admin'
-            ? Post::query()
-            : Post::where('is_visible', true);
-
-        // If the user is authenticated, retrieve posts based on the selected category or get all posts
-        if (Auth::check()) {
-            $posts = $categoryId
-                ? $query->whereHas('categories', function ($query) use ($categoryId) {
-                    $query->where('categories.id', $categoryId);
-                })->get()
-                : $query->get();
-        } else {
-            // If the user is not authenticated, only retrieve visible posts
-            $posts = Post::where('is_visible', true)->get();
+        // If the user is authenticated and a category is selected, filter by category
+        if (Auth::check() && $categoryId) {
+            $query->whereHas('categories', function ($query) use ($categoryId) {
+                $query->where('categories.id', $categoryId);
+            });
         }
 
-        return view('posts.index', compact('posts', 'categories', 'categoryId'));
+        // Retrieve posts based on the constructed query
+        $posts = $query->with('categories')->get();
+
+
+        // Pass $categories directly to partial view
+        return view('posts.index', compact('posts', 'categoryId', 'categories'));
     }
+
+
+    public function search(Request $request)
+    {
+        $categories = Category::all();
+        // Get the search value from the request
+        $search = $request->input('search');
+        $searchCategories = $request->input('searchCategory');
+        $posts = null;
+
+        // Search in the title and body columns from the artworks table
+        if (!$search == null) {
+            $posts = Post::query()
+                ->where('title', 'LIKE', "%{$search}%")
+                ->orWhere('detail', 'LIKE', "%{$search}%")
+                ->where('is_visible', true) // Add the visibility check for the search query
+                ->get();
+        }
+
+        if (!$searchCategories == null) {
+            $i = 0;
+            foreach ($searchCategories as $searchCategory) {
+                if ($i === 0) {
+                    $i++;
+                    $posts = Post::query()
+                        ->where('title', 'LIKE', "%{$search}%")
+                        ->whereHas('categories', function ($query) use ($searchCategory) {
+                            $query->where('categories.id', $searchCategory);
+                        })
+                        ->where('is_visible', true) // Add the visibility check for the search query
+                        ->get();
+                } elseif ($i > 0) {
+                    $posts = Post::query()
+                        ->where('title', 'LIKE', "%{$search}%")
+                        ->whereHas('categories', function ($query) use ($searchCategory) {
+                            $query->where('categories.id', $searchCategory);
+                        })
+                        ->where('is_visible', true) // Add the visibility check for the search query
+                        ->get();
+                }
+            }
+        }
+        // Return the search view with the results compacted
+        return view('posts.index', compact('posts', 'categories'));
+    }
+
 
 
     public function create()
@@ -56,33 +95,37 @@ class PostController extends Controller
         return view('posts.create', compact('categories'));
     }
 
-
     public function store(Request $request)
     {
-        // Validate and store post
-        // Implement your validation logic based on your requirements
-        $request->validate([
+        // Validate the request data
+        $validatedData = $request->validate([
             'title' => 'bail|required|max:255',
             'detail' => 'bail|required|max:255',
             'image' => 'required|mimes:jpg,jpeg,png|max:4096',
-            'category_id' => 'required|exists:categories,id',
+            'category_id' => 'bail|required',
+            'category_id.*' => 'bail|required|max:255|exists:categories,id'
             // Add other fields as needed
         ]);
 
-        // Assuming the 'image' field is a file, handle file upload
+        // Handle file upload
         $imagePath = $request->file('image')->store('post_images', 'public');
 
-        // Create a new Post instance with the validated data
-        $post = new Post([
-            'title' => $request->input('title'),
-            'detail' => $request->input('detail'),
-            'image' => $imagePath,
-            'category_id' => $request->input('category_id'),
-            'is_visible' => $request->has('is_visible'),
-        ]);
+        // Merge additional data, including user_id and image path
+        $post = new Post(request(['title',
+            'detail',
+            'image',
+            'category_id',]));
 
-        // Save the post to the database
+        $post->user_id = Auth::user()->id;
+
+        $post->user_name = Auth::user()->id;
+
+        $post->image = $imagePath;
+
         $post->save();
+
+        // Attach the selected category to the post
+        $post->categories()->attach($validatedData['category_id']);
 
         // Redirect back to the main page with a success message
         return redirect()->route('posts.index')->with('success', 'Post created successfully');
@@ -91,8 +134,16 @@ class PostController extends Controller
     public function edit($id)
     {
         $post = Post::findOrFail($id);
-        $categories = Category::all();
-        return view('posts.edit', compact('post', 'categories'));
+//
+//        // Check if the currently authenticated user is the owner of the post
+        if ($post->user_id == Auth::user()->id) {
+
+            $categories = Category::all();
+
+            return view('posts.edit', compact('post', 'categories'));
+        }
+
+        return redirect()->route('posts.index')->with('error', 'Unauthorized access');
     }
 
     public function update(Request $request, $id)
@@ -101,10 +152,10 @@ class PostController extends Controller
         $request->validate([
             'title' => 'nullable|max:255',
             'detail' => 'nullable|max:255',
-            'image' => 'nullable|mimes:jpg,jpeg,png|max:4096', // Allow image to be nullable
+            'image' => 'nullable|mimes:jpg,jpeg,png|max:4096',
             'is_visible' => 'boolean',
-            'category_id' => 'required|exists:categories,id',
-            // Add other fields as needed
+            'category_id' => 'bail|required',
+            'category_id.*' => 'bail|required|max:255|exists:categories,id'
         ]);
 
         $post = Post::findOrFail($id);
@@ -118,11 +169,14 @@ class PostController extends Controller
 
         // Update other fields
         $post->title = $request->input('title');
-        $post->category_id = $request->input('category_id');
+        $post->detail = $request->input('detail');
         $post->is_visible = $request->has('is_visible');
 
         // Save the changes
         $post->save();
+
+        // Update the category
+        $post->categories()->sync($request->input('category_id'));
 
         return redirect()->route('posts.index')->with('success', 'Post updated successfully');
     }
@@ -130,11 +184,12 @@ class PostController extends Controller
     public function show($postId)
     {
         $post = Post::findOrFail($postId);
+        $category = $post->category; // Assuming 'category' is a relationship on your Post model
 
         // Track post view
         $this->trackPostView($post);
 
-        return view('posts.show', compact('post'));
+        return view('posts.show', compact('post', 'category'));
     }
 
     private function trackPostView($post)
@@ -163,29 +218,17 @@ class PostController extends Controller
     public function destroy($id)
     {
         $post = Post::findOrFail($id);
-        $post->delete();
 
-        return redirect()->route('posts.index')->with('success', 'Post deleted successfully');
-    }
+//        // Check if the currently authenticated user is the owner of the post
+        if (Auth::user()->id == $post->user_id) {
 
-    public function search(Request $request)
-    {
-        // Retrieve the search query from the request
-        $query = $request->input('q');
+            $post->delete();
 
-        // Perform the search logic (adjust this based on your requirements)
-        $results = Post::where('title', 'like', "%{$query}%")->get();
+            return redirect()->route('posts.index')->with('success', 'Post deleted successfully');
+        }
 
-        // Pass the results to the view
-        return view('posts.search', compact('results'));
-    }
+        return redirect()->route('posts.index')->with('error', 'Unauthorized access');
 
-    public function filterByCategory(Category $category)
-    {
-        $posts = $category->posts()->where('is_visible', true)->get();
-
-        // You may want to load the view or return a JSON response, depending on your requirements
-        return view('posts.index', compact('posts'));
     }
